requirements.txt
requests~=2.32.3
beautifulsoup4~=4.13.3
pandas~=2.2.3
python-dotenv~=1.0.1
duckdb~=1.2.1
matplotlib~=3.10.1
seaborn~=0.13.2
numpy~=2.2.3
ipython~=9.0.2